feat: returned max and min values for homography transform
