+++ main.py
@@ -2,6 +2,7 @@ from detection.player_detector import PlayerDetector
 from modeling.court_drawer import CourtDrawer
 from pre_processing.image_loader import ImageLoader
 from homography.homography_calculator import HomographyCalculator
+from get_field_outline import detect_court_boundary
 import os
 import re
 
@@ -23,11 +24,10 @@ def main():
         img = ImageLoader(img_path)
         detector = PlayerDetector(img)
         drawer = CourtDrawer()
-        calc = HomographyCalculator()
         court_corners = drawer.right_bounds
-        _, encoded_img = img.load_and_encode_image()
+        image = img.load_image()
         player_positions = detector.detect_players_with_mask_rcnn(image_path=img_path)
-        camera_view_corners = calc.fetch_points_for_homography(encoded_img)
+        camera_view_corners = detect_court_boundary(image)
         drawer.plot_transformed_positions(
             player_positions, camera_view_corners, court_corners
         )
@@ -41,7 +41,7 @@ def vectors():
     calc = HomographyCalculator()
     _, encoded_img = img.load_and_encode_image()
     player_positions = detector.detect_players_with_mask_rcnn(image_path=img_path)
-    camera_view_corners = calc.fetch_points_for_homography(encoded_img)
+    camera_view_corners = detect_court_boundary(encoded_img)
     left_corner = camera_view_corners[0]
     vectors = calc.calculate_vectors(player_positions, left_corner)
     drawer.plot_vectors(vectors)
